requirements.txt
aiohttp==3.8.3
python-telegram-bot==20.0a4